+++ src/atom/categories.js
@@ -1,0 +1,5 @@
+import { atom } from "recoil";
+
+export const entireCategory = atom({ key: "entireCategory", default: [] });
+export const userCategory = atom({ key: "userCategory", default: 0 });
+export const userCategoryName = atom({ key: "userCategoryName", default: "" });
