+++ src/atom/sortBy.js
@@ -1,0 +1,4 @@
+import { atom } from "recoil";
+
+// eslint-disable-next-line import/prefer-default-export
+export const sortBy = atom({ key: "sortBy", default: "" });
