+++ src/css/Footer.css
@@ -4,7 +4,7 @@
   width: 100%;
   min-width: 1440px;
   height: 380px;
-  margin-top: 300px;
+  /* margin-top: 300px; */
 }
 
 .footer__text {
