+++ src/css/WishBook.css
@@ -1,0 +1,32 @@
+.wish-book {
+  flex: none;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.wish-book__icon {
+  width: 48px;
+  height: 48px;
+}
+
+.wish-book__title {
+  margin-top: 44px;
+}
+
+.wish-book__subtitle {
+  margin-top: 16px;
+  margin-bottom: 48px;
+}
+
+.wish-book__button {
+  margin: 0;
+  border: 0;
+  width: 265px;
+  height: 48px;
+  padding: 9px 36px 10px;
+  background: none;
+  border-radius: 24px;
+  background-color: #e63737;
+}
